# Optimize `String#valid_encoding?`