This PR implements one of the algorithms from #11873.

The loop can be made even faster by unrolling, which is not done here yet. [Here is a benchmark](https://gist.github.com/HertzDevil/6022606e107bdfe0c1035858d5f54577) comparing 4 different implementations of `String#valid_encoding?`:

* `old`: The existing `Char::Reader`-based loop.
* `old inlined`: As above, but with [this modification](https://github.com/crystal-lang/crystal/issues/11873#issuecomment-1059956869) applied to `Char::Reader#next_char`.
* `new`: This PR.
* `new unrolled`: Adds the following unrolling code, where `UNROLL = 64` for the benchmark:
  ```crystal
  while s + UNROLL <= e
    {% for i in 0...UNROLL %}
      state = table[s[{{ i }}]].unsafe_shr(state & 0x3F)
    {% end %}
    return false if state & 0x3F == 6
    s += UNROLL
  end
  ```
  The `return` statement must be outside the inner loop, otherwise the unrolling would become only marginally faster than this PR.

The results are:

```
         old 936.05  (  1.07ms) (± 2.37%)  0.0B/op   2.93× slower
 old inlined   1.57k (638.75µs) (± 3.85%)  0.0B/op   1.75× slower
         new   1.88k (532.38µs) (± 7.53%)  0.0B/op   1.46× slower
new unrolled   2.75k (364.02µs) (± 3.42%)  0.0B/op        fastest
```

With `--mcpu=native` this becomes:

```
         old 931.93  (  1.07ms) (± 3.02%)  0.0B/op   5.19× slower
 old inlined   1.57k (635.39µs) (± 3.11%)  0.0B/op   3.07× slower
         new   2.18k (458.16µs) (± 6.94%)  0.0B/op   2.22× slower
new unrolled   4.84k (206.70µs) (± 3.54%)  0.0B/op        fastest
```

As explained in the linked issue, this improvement relies on the `shrx` instruction from the [BMI2](https://en.wikipedia.org/wiki/X86_Bit_manipulation_instruction_set#BMI2_(Bit_Manipulation_Instruction_Set_2)) extension. The `Char::Reader` implementations are barely affected.

Note that if the first byte of a sufficiently long string is already invalid, the inner loop must still consume the rest of the `UNROLL`-byte chunk before the method returns. This is what happens if the benchmark had `s = "\x80" * N` instead:

```
         old  78.94M ( 12.67ns) (± 6.46%)  0.0B/op   3.56× slower
 old inlined 280.90M (  3.56ns) (± 5.88%)  0.0B/op        fastest
         new 275.24M (  3.63ns) (± 8.38%)  0.0B/op   1.02× slower
new unrolled  20.22M ( 49.45ns) (± 4.52%)  0.0B/op  13.89× slower
```

Because of this I am not sure if we should unroll the loop here.