@HertzDevil 2022-06-22
Doing the byte loop before the unrolled loop does not matter, because if the first invalid byte occurs within the first unrolled chunk, then the method now visits `bytes.size % UNROLL + UNROLL` bytes instead of `UNROLL` bytes; the average iteration count would ultimately be the same.

Moving out the `return` in the byte loop did little for valid strings, but now the method must also iterate through all bytes for invalid strings shorter than `UNROLL` bytes. So I think the benchmark's version is good enough.

What I wonder most is whether unrolling still gives a performance boost on AArch64.