@straight-shoota 2022-08-01
@HertzDevil The implementation still needs to be integrated into `valid_encoding?`, right?