@HertzDevil 2022-08-13
Crystal 1.0.0 cannot handle `UInt64`s inside the `dfa_state` macro until #10713.

:thinking: